FIX: Convert to required type
Pandas has breaking changes in 1.4.0 which requires us to explicitly
convert the values.

+++ src/xtgeo/xyz/_xyz_io.py
@@ -330,7 +330,9 @@ def _convert_idbased_xyz(self, df):
 
     idgroups = df.groupby(self._pname)
 
-    newdf = pd.DataFrame(columns=[self._xname, self._yname, self._zname])
+    newdf = pd.DataFrame(
+        columns=[self._xname, self._yname, self._zname], dtype="float64"
+    )
     udef = pd.DataFrame(
         [[999.0, 999.0, 999.0]], columns=[self._xname, self._yname, self._zname]
     )
